Move strategy files into a strategy/ subfolder
bot-strategy, forced-win and search (with their specs) form the bot's
brain and only depend on geometry/helpers; grouping them separates the
strategy from the board/game UI files.
