+++ src/components/games/triangle-circle-game/strategy/search.spec.ts
@@ -1,8 +1,8 @@
-import { EDGES, TRIANGLES, TRIANGLE_COUNT } from './geometry';
+import { EDGES, TRIANGLES, TRIANGLE_COUNT } from '../geometry';
 import {
   type Board, LINE, CIRCLE,
   generateStartBoard, applyShade
-} from './helpers';
+} from '../helpers';
 import { evaluatePosition } from './search';
 
 const otherEdge = (t: number, notEdge: number) => TRIANGLES[t].edgeIds.find(e => e !== notEdge)!;
